pyproject.toml
[tool.poetry]
name = "diary-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.2.1"
langchain-chroma = "^0.1.1"
langchain-openai = "^0.1.7"
langchain-google-genai = "^1.0.5"
python-docx = "^1.1.2"
azure-identity = "^1.16.0"
azure-search-documents = "^11.4.0"
langchain-community = "^0.2.1"
unstructured = "^0.14.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
